+++ Jarvis.Framework.Kernel/ProjectionEngine/CollectionWrapper.cs
@@ -139,6 +139,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine
 
             OnSave(model, e, CollectionWrapperOperationType.Insert);
             model.Version = 1;
+            model.AggregateVersion = e.Version;
             model.LastModified = e.CommitStamp;
 
             model.AddEvent(e.MessageId);
@@ -207,6 +208,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine
             OnSave(model, e, CollectionWrapperOperationType.Update);
             var orignalVersion = model.Version;
             model.Version++;
+            model.AggregateVersion = e.Version;
             model.LastModified = e.CommitStamp;
             model.AddEvent(e.MessageId);
             HandlePollableReadModel(e, model);